Converts AsyncPoco.Tests project to use NUnit testing framework
The previous code used a proprietary PetaTest.cs library to execute all of the tests. By switching to NUnit, we can retain all of the existing code, but enjoy a more flexible testing framework. I've also included support for NUnitLite so that we can continue running it from a CLI, but we can also run and debug one or more tests from within Visual Studio.

I had originally tried to convert the code to xUnit, but I found it to be very difficult to get the tests working reliably. There is a lot of extra code that needs to be written to achieve the same kinds of testing that we can achieve for free with NUnit.

Other changes:
* Removed the csproj.user file
* Removed the PetaTest.cs file
* Renamed SqlBuilder to SqlBuilderTests
* Renamed Tests to DatabaseTests
* Updated the .gitignore file to exclude *.user files to avoid accidentally committing them in the future.